PyQt5_GUI/requirements.txt
PyQt5-sip<13,>=12.11
PyQt5-Qt5>=5.15.0